[Engine] Fixed coordinates send to shader for texture position

+++ illagameengine-libgdx/src/main/java/org/illarion/engine/backend/gdx/GdxGrayScaleEffect.java
@@ -62,10 +62,10 @@ class GdxGrayScaleEffect implements GrayScaleEffect, GdxSceneEffect, GdxTextureE
     }
 
     @Override
-    public void setTopLeftCoordinate(int x, int y) {
+    public void setTopLeftCoordinate(float x, float y) {
     }
 
     @Override
-    public void setBottomRightCoordinate(int x, int y) {
+    public void setBottomRightCoordinate(float x, float y) {
     }
 }
