+++ illagameengine-libgdx/src/main/java/org/illarion/engine/backend/gdx/GdxMiniMapEffect.java
@@ -82,11 +82,11 @@ class GdxMiniMapEffect implements MiniMapEffect, GdxTextureEffect {
     }
 
     @Override
-    public void setTopLeftCoordinate(int x, int y) {
+    public void setTopLeftCoordinate(float x, float y) {
     }
 
     @Override
-    public void setBottomRightCoordinate(int x, int y) {
+    public void setBottomRightCoordinate(float x, float y) {
     }
 
     @Override
