+++ illagameengine-libgdx/src/main/java/org/illarion/engine/backend/gdx/GdxTileLightEffect.java
@@ -170,12 +170,12 @@ public class GdxTileLightEffect implements TileLightEffect, GdxTextureEffect {
     }
 
     @Override
-    public void setTopLeftCoordinate(int x, int y) {
+    public void setTopLeftCoordinate(float x, float y) {
         topLeftCoord.set(x, y);
     }
 
     @Override
-    public void setBottomRightCoordinate(int x, int y) {
+    public void setBottomRightCoordinate(float x, float y) {
         bottomRightCoord.set(x, y);
     }
 }
